Introduce metrics to the system
There are currently two metrics implementations that most of the crates use:
* [OpenTelemetry](https://docs.rs/opentelemetry/latest/opentelemetry/#metrics)
* [Metrics](https://docs.rs/metrics/0.19.0/metrics/)

Both provide similar API that allows creating counters, gauges and histograms.
However, there seem to be a warning that OpenTelemetry implementation is not stable.
Also, OpenTelemetry crate is more than just metrics, it also provides tracing API,
which may shadow `tracing` if pulled in scope.

Because of these reasons, this change suggests to use `Metrics` crate. However, it is
likely that we would need to start using OpenTelemetry at some point because it
provides integration with many open-source and third-party metric/trace collectors
and visualization tools. I don't believe we will have to rewrite IPA implementation
to do that, because it is possible to integrate tracing with OpenTelemetry using
[tracing-opentelemetry](https://docs.rs/tracing-opentelemetry/0.17.4/tracing_opentelemetry/)
crate and integrate `Metrics` by creating a `Recorder` that just dumps everything
to `OpenTelemetry` sink

* Add `Metrics` crate as dependency
* Modify tower tracing layer to emit `requests.received` metric
* Unit tests for the above
* A toy implementation of metrics sinkkk for CLI that dumps everything to stderr